Release 1.1.1 - Mejora el historial de eventos con hora separada, dirección de viaje, vigencia de transferencia y montos más robustos. - Añade soporte para Mi Macro Aeropuerto, nuevas estaciones y más identificadores de rutas/operadores de Mi Movilidad. - Corrige el parseo y respaldo de IdSAMDistribución usando valores largos/unsigned. - Agrega migración Room v7 para preservar IDs SAM existentes. - Refactoriza utilidades de fechas compactas/BCD y amplía pruebas de parser, backup y mappers. - Actualiza versionCode a 26, versionName a 1.1.1 y compileSdk a 37.

+++ app/src/main/java/org/nodocentral/miviaje/data/room/ProductEntity.java
@@ -38,7 +38,7 @@ public class ProductEntity {
     public int distributorNetworkId;
     public short distributorCompanyId;
     public LocalDateTime distributionDateTime;
-    public int distributionSamId;
+    public long distributionSamId;
     public short distributingDeviceId;
     public LocalDateTime validFrom;
     public LocalDateTime validTo;
@@ -61,4 +61,3 @@ public class ProductEntity {
     public int lastDebitRouteStationId;
     public int lastDebitDeviceId;
 }
-
